fix: main fn separation
