+++ core/src/route/get.rs
@@ -1,0 +1,5 @@
+pub mod article;
+pub mod category;
+pub mod error;
+pub mod index;
+pub mod tag;
